Kext can be loaded
This is the initial commit for WN722N-V1.
It has been developed and tested only in Catalina using Xcode 11.3, it loads fine, it writes the firmware to the device and currently writes a fixed MAC address(TODO), the one my device has. It creates and publishes the interface but it is detected as an ethernet device(TODO). No further work has been done.